Add support for fractional scaling
When the logical layout mode is used, allow configuring the scaling to
be non-integer. Todo: expand on this

+++ src/tests/monitor-store-unit-tests.c
@@ -51,7 +51,7 @@ typedef struct _MonitorTestCaseMonitor
 typedef struct _MonitorTestCaseLogicalMonitor
 {
   MetaRectangle layout;
-  int scale;
+  float scale;
   MetaMonitorTransform transform;
   gboolean is_primary;
   gboolean is_presentation;
